Own checkpoint API models in core
