+++ Sources/Checkpoints/Checkpoint.swift
@@ -15,6 +15,8 @@
 import CoreFoundation
 import Foundation
 
+// swiftlint:disable file_length
+
 /// A custom value supplied when a checkpoint is hit.
 @_spi(Internal) public enum CheckpointValue: Equatable, Hashable, Sendable {
 
@@ -179,6 +181,251 @@ extension CheckpointInfo: CustomStringConvertible {
 
 }
 
+/// The reason no experience was served for a checkpoint.
+@_spi(Internal) public struct CheckpointNoActionReason: Equatable, Hashable, Sendable {
+
+    /// The value identifying the reason.
+    public let value: String
+
+    /// No targeting rule matched.
+    public static let noMatch = Self(value: "NO_MATCH")
+
+    /// The customer was assigned to a holdout.
+    public static let holdout = Self(value: "HOLDOUT")
+
+    /// The customer reached the configured frequency cap.
+    public static let frequencyCapped = Self(value: "FREQUENCY_CAPPED")
+
+    /// Checkpoint configuration could not be loaded.
+    public static let configurationUnavailable = Self(value: "CONFIGURATION_UNAVAILABLE")
+
+    /// Checkpoints are disabled.
+    public static let disabled = Self(value: "DISABLED")
+
+    init(value: String) {
+        self.value = value
+    }
+
+}
+
+extension CheckpointNoActionReason: CustomStringConvertible {
+
+    /// The raw reason value.
+    public var description: String { return self.value }
+
+}
+
+/// Base class for the result of hitting a checkpoint.
+///
+/// Inspect the concrete subclass to determine the outcome. This class cannot be initialized or subclassed outside
+/// the SDK. New result subclasses can be added without breaking application code that inspects the concrete type.
+@_spi(Internal) public class CheckpointResult: CustomStringConvertible, Equatable {
+
+    /// Information about the checkpoint that produced this result.
+    public let checkpoint: CheckpointInfo
+
+    fileprivate init(checkpoint: CheckpointInfo) {
+        self.checkpoint = checkpoint
+    }
+
+    /// A debug description of the checkpoint result.
+    public var description: String {
+        return "CheckpointResult(checkpoint=\(self.checkpoint))"
+    }
+
+    /// Returns whether two checkpoint results have the same concrete type and associated values.
+    public static func == (lhs: CheckpointResult, rhs: CheckpointResult) -> Bool {
+        return lhs.isEqual(to: rhs)
+    }
+
+    fileprivate func isEqual(to other: CheckpointResult) -> Bool {
+        return type(of: self) == type(of: other) && self.checkpoint == other.checkpoint
+    }
+
+}
+
+/// A checkpoint-triggered paywall was presented and finished.
+@_spi(Internal) public final class CheckpointPaywallPresentedResult: CheckpointResult {
+
+    /// The terminal outcome of the presented paywall.
+    public let paywallOutcome: CheckpointPaywallOutcome
+
+    fileprivate init(checkpoint: CheckpointInfo, paywallOutcome: CheckpointPaywallOutcome) {
+        self.paywallOutcome = paywallOutcome
+        super.init(checkpoint: checkpoint)
+    }
+
+    /// A debug description of the presented-paywall result.
+    public override var description: String {
+        return "PaywallPresented(checkpoint=\(self.checkpoint), paywallOutcome=\(self.paywallOutcome))"
+    }
+
+    fileprivate override func isEqual(to other: CheckpointResult) -> Bool {
+        guard let other = other as? CheckpointPaywallPresentedResult else { return false }
+        return self.checkpoint == other.checkpoint && self.paywallOutcome == other.paywallOutcome
+    }
+
+}
+
+/// Nothing was served for a checkpoint.
+@_spi(Internal) public final class CheckpointNoActionResult: CheckpointResult {
+
+    /// The reason no experience was served.
+    public let reason: CheckpointNoActionReason
+
+    fileprivate init(checkpoint: CheckpointInfo, reason: CheckpointNoActionReason) {
+        self.reason = reason
+        super.init(checkpoint: checkpoint)
+    }
+
+    /// A debug description of the no-action result.
+    public override var description: String {
+        return "NoAction(checkpoint=\(self.checkpoint), reason=\(self.reason))"
+    }
+
+    fileprivate override func isEqual(to other: CheckpointResult) -> Bool {
+        guard let other = other as? CheckpointNoActionResult else { return false }
+        return self.checkpoint == other.checkpoint && self.reason == other.reason
+    }
+
+}
+
+/// Base class for the terminal result of a checkpoint-presented paywall.
+///
+/// Inspect the concrete subclass to determine the outcome. This class cannot be initialized or subclassed outside
+/// the SDK. New outcome subclasses can be added without breaking application code that inspects the concrete type.
+@_spi(Internal) public class CheckpointPaywallOutcome: CustomStringConvertible, Equatable {
+
+    fileprivate init() {}
+
+    /// A debug description of the paywall outcome.
+    public var description: String {
+        return "CheckpointPaywallOutcome"
+    }
+
+    /// Returns whether two paywall outcomes have the same concrete type and associated values.
+    public static func == (lhs: CheckpointPaywallOutcome, rhs: CheckpointPaywallOutcome) -> Bool {
+        return lhs.isEqual(to: rhs)
+    }
+
+    fileprivate func isEqual(to other: CheckpointPaywallOutcome) -> Bool {
+        return type(of: self) == type(of: other)
+    }
+
+}
+
+/// The customer dismissed the paywall.
+@_spi(Internal) public final class CheckpointPaywallDismissedOutcome: CheckpointPaywallOutcome {
+
+    /// The dismissed paywall outcome.
+    public static let shared = CheckpointPaywallDismissedOutcome()
+
+    private override init() {
+        super.init()
+    }
+
+    /// A debug description of the dismissed outcome.
+    public override var description: String { return "Dismissed" }
+
+}
+
+/// The customer completed a purchase.
+@_spi(Internal) public final class CheckpointPaywallPurchasedOutcome: CheckpointPaywallOutcome {
+
+    /// Customer information after the completed purchase.
+    public let customerInfo: CustomerInfo
+
+    /// Creates a purchased outcome with the latest customer information.
+    public init(customerInfo: CustomerInfo) {
+        self.customerInfo = customerInfo
+        super.init()
+    }
+
+    /// A debug description of the purchased outcome.
+    public override var description: String { return "Purchased" }
+
+    fileprivate override func isEqual(to other: CheckpointPaywallOutcome) -> Bool {
+        guard let other = other as? CheckpointPaywallPurchasedOutcome else { return false }
+        return self.customerInfo.isEqual(other.customerInfo)
+    }
+
+}
+
+/// The customer restored purchases.
+@_spi(Internal) public final class CheckpointPaywallRestoredOutcome: CheckpointPaywallOutcome {
+
+    /// Customer information after restoring purchases.
+    public let customerInfo: CustomerInfo
+
+    /// Creates a restored outcome with the latest customer information.
+    public init(customerInfo: CustomerInfo) {
+        self.customerInfo = customerInfo
+        super.init()
+    }
+
+    /// A debug description of the restored outcome.
+    public override var description: String { return "Restored" }
+
+    fileprivate override func isEqual(to other: CheckpointPaywallOutcome) -> Bool {
+        guard let other = other as? CheckpointPaywallRestoredOutcome else { return false }
+        return self.customerInfo.isEqual(other.customerInfo)
+    }
+
+}
+
+/// The paywall ended with an error.
+@_spi(Internal) public final class CheckpointPaywallErrorOutcome: CheckpointPaywallOutcome {
+
+    /// The error that ended the checkpoint experience.
+    public let error: PublicError
+
+    /// Creates a paywall error outcome.
+    public init(error: PublicError) {
+        self.error = error
+        super.init()
+    }
+
+    /// A debug description of the error outcome.
+    public override var description: String {
+        return "Error(error=\(self.error))"
+    }
+
+    fileprivate override func isEqual(to other: CheckpointPaywallOutcome) -> Bool {
+        guard let other = other as? CheckpointPaywallErrorOutcome else { return false }
+        return self.error.isEqual(other.error)
+    }
+
+}
+
+/// Global listener for checkpoint activity.
+///
+/// All methods are called on the main thread.
+@_spi(Internal) public protocol CheckpointListener: AnyObject {
+
+    /// A checkpoint was hit, before evaluation.
+    func onCheckpointHit(_ checkpoint: CheckpointInfo)
+
+    /// The checkpoint was resolved and the outcome was returned.
+    func onCheckpointResolved(_ checkpoint: CheckpointInfo, result: CheckpointResult)
+
+    /// A checkpoint-presented paywall finished.
+    func onCheckpointPaywallFinished(_ checkpoint: CheckpointInfo, outcome: CheckpointPaywallOutcome)
+
+}
+
+@_spi(Internal) public extension CheckpointListener {
+
+    /// Default no-op implementation.
+    func onCheckpointHit(_ checkpoint: CheckpointInfo) {}
+
+    /// Default no-op implementation.
+    func onCheckpointResolved(_ checkpoint: CheckpointInfo, result: CheckpointResult) {}
+
+    /// Default no-op implementation.
+    func onCheckpointPaywallFinished(_ checkpoint: CheckpointInfo, outcome: CheckpointPaywallOutcome) {}
+
+}
+
 /// Internal representation of why a checkpoint did not run anything.
 @_spi(Internal) public struct CheckpointEngineNoActionReason: Equatable, Sendable {
 
@@ -281,5 +528,43 @@ extension CheckpointInfo: CustomStringConvertible {
         _ checkpoint: CheckpointInfo,
         outcome: CheckpointEnginePaywallOutcome
     )
+
+}
+
+@_spi(Internal) public extension CheckpointResult {
+
+    /// Converts the internal engine result into its checkpoint API representation.
+    static func from(_ result: CheckpointEngineResult) -> CheckpointResult {
+        switch result {
+        case let .paywallPresented(checkpoint, outcome):
+            return CheckpointPaywallPresentedResult(
+                checkpoint: checkpoint,
+                paywallOutcome: CheckpointPaywallOutcome.from(outcome)
+            )
+        case let .noAction(checkpoint, reason):
+            return CheckpointNoActionResult(
+                checkpoint: checkpoint,
+                reason: CheckpointNoActionReason(value: reason.value)
+            )
+        }
+    }
+
+}
+
+@_spi(Internal) public extension CheckpointPaywallOutcome {
+
+    /// Converts the internal engine outcome into its checkpoint API representation.
+    static func from(_ outcome: CheckpointEnginePaywallOutcome) -> CheckpointPaywallOutcome {
+        switch outcome {
+        case .dismissed:
+            return CheckpointPaywallDismissedOutcome.shared
+        case let .purchased(customerInfo):
+            return CheckpointPaywallPurchasedOutcome(customerInfo: customerInfo)
+        case let .restored(customerInfo):
+            return CheckpointPaywallRestoredOutcome(customerInfo: customerInfo)
+        case let .error(error):
+            return CheckpointPaywallErrorOutcome(error: error)
+        }
+    }
 
 }
